fix: emitting struct in event just yields hash of data (#10)
Added all `Subscription` fields to both events. I don't think it is
necessary to include all parameters in the `Redeem` event, however, kept
this as is for now.

We should decide what fields to remove, I propose removing `amount` and
`frequency`. Rationale:
- shouldn't be changeable after subscription creation, whereas we may
add functionality for the subscription recipient to change: the
recipient address & whether they require a trusted path.